Create diff files for IonizationState & IonizationStates
